requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.13
openai==1.12.0
chromadb==0.4.22
pypdf2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
faiss-cpu==1.7.4
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
sentence-transformers==2.2.2